backend/requirements.txt
Flask
flask-cors
supabase
python-dotenv
httpcore>=1.0.8
httpx>=0.27.0
